Add search for groups and users along with many group routes

+++ Backend/server/controllers/group.controller.js
@@ -34,15 +34,14 @@ const createGroup = catchAsyncError(async (req, res, next) => {
 				);
 
 				res.json({
-					message: "Group Created",
 					status: true,
 					send: groupData,
 				});
 			})
 			.catch((error) => {
 				res.json({
-					message: `Group Creation Failed! ${error.message}`,
 					status: false,
+					message: `Group Creation Failed! ${error.message}`,
 				});
 			});
 	} catch (error) {
@@ -50,10 +49,75 @@ const createGroup = catchAsyncError(async (req, res, next) => {
 	}
 });
 
+const findUser = catchAsyncError(async (req, res, next) => {
+	try {
+		const username = req.query.username;
+		const user = await User.findOne({ username })
+			.select("username avatar")
+			.populate("avatar", "image -_id");
+		console.log(user);
+		if (user) {
+			res.json({
+				status: true,
+				username: user.username,
+				avatar: user.avatar.image,
+				id: user._id,
+			});
+		} else {
+			res.json({
+				status: false,
+				message: "User Not Found!",
+			});
+		}
+	} catch (error) {
+		return next(new ErrorHandler(error.message, 400));
+	}
+});
+
+const addMembers = catchAsyncError(async (req, res, next) => {
+	try {
+		// add group to user doc
+		const groupName = req.params["group"];
+		const { members } = req.body;
+		const group = await Group.findOneAndUpdate(
+			{ groupName },
+			{ $push: { members: { $each: [...members] } } }
+		);
+		for (const member of members) {
+			await User.findOneAndUpdate(
+				{ _id: member },
+				{ $push: { groups: group._id } }
+			);
+		}
+
+		res.json({
+			status: true,
+			message: "Successfully Added New Members",
+		});
+	} catch (error) {
+		return next(new ErrorHandler(error.message, 400));
+	}
+});
+
+const getGroupDetails = catchAsyncError(async (req, res, next) => {
+	try {
+		const groupName = req.params["group"];
+		const group = await Group.findOne({ groupName });
+		const memberList = await group.getGroupDetails();
+
+		res.json({
+			status: true,
+			groupName: group.groupName,
+			description: group.description,
+			members: memberList,
+		});
+	} catch (error) {
+		return next(new ErrorHandler(error.message, 400));
+	}
+});
+
 const updateAvatar = catchAsyncError(async (req, res, next) => {
 	try {
-		const { username } = req.user;
-
 		const allowedTypes = ["image/jpeg", "image/png"];
 		if (!allowedTypes.includes(req.file.mimetype)) {
 			return next(
@@ -70,10 +134,12 @@ const updateAvatar = catchAsyncError(async (req, res, next) => {
 				new ErrorHandler("Image Size Too Large! (Max Size <5MB)", 400)
 			);
 		}
+		const groupName = req.params["group"];
 
-		const group = await Group.findOne({ groupName });
-		await Image.deleteOne({ _id: group.avatar });
-
+		const group = await Group.findOne({ groupName }).select("avatar");
+		if (group) {
+			await Image.deleteOne({ _id: group.avatar });
+		}
 		const base64Image = req.file.buffer.toString("base64");
 		const avatar = await Image.create({
 			name: req.file.originalname,
@@ -82,12 +148,18 @@ const updateAvatar = catchAsyncError(async (req, res, next) => {
 		await Group.findOneAndUpdate({ groupName }, { avatar: avatar._id });
 
 		res.json({
-			message: "Avatar Updated Successfully",
 			status: true,
+			message: "Avatar Updated Successfully",
 		});
 	} catch (error) {
 		return next(new ErrorHandler(error.message, 400));
 	}
 });
 
-module.exports = { createGroup, updateAvatar };
+module.exports = {
+	createGroup,
+	updateAvatar,
+	findUser,
+	addMembers,
+	getGroupDetails,
+};
